requirements.txt
asgiref==3.4.1
Django==3.2.9
django-quill-editor==0.1.22
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-taggit==2.0.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
whitenoise==5.3.0
